# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Enterprise Device Management Service for POS systems, printers, and payment terminals",
        "title": "Device Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Device Service API Support",
            "email": "support@deviceservice.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8084",
    "basePath": "/api/v1",
    "paths": {
...
        "/discovery/auto-setup": {
            "post": {
                "description": "Automatically register and setup discovered devices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discovery"
                ],
                "summary": "Auto-setup devices",
                "parameters": [
                    {
                        "description": "Auto-setup request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AutoSetupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auto-setup completed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.AutoSetupResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Auto-setup failed",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/discovery/capabilities/{brand}/{type}": {
            "get": {
                "description": "Get capabilities for a specific brand and device type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discovery"
                ],
                "summary": "Get device capabilities",
                "parameters": [
                    {
                        "enum": [
                            "EPSON",
                            "STAR",
                            "INGENICO",
                            "PAX",
                            "CITIZEN",
                            "BIXOLON",
                            "VERIFONE",
                            "GENERIC"
                        ],
                        "type": "string",
                        "description": "Device brand",
                        "name": "brand",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "POS",
                            "PRINTER",
                            "SCANNER",
                            "CASH_REGISTER",
                            "CASH_DRAWER",
                            "DISPLAY"
                        ],
                        "type": "string",
                        "description": "Device type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Capabilities retrieved",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "brand": {
                                                    "type": "string"
                                                },
                                                "capabilities": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "device_type": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Device not supported",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/discovery/scan": {
            "get": {
                "description": "Scan for available devices on serial, USB, TCP, or Bluetooth connections",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discovery"
                ],
                "summary": "Scan for devices",
                "parameters": [
                    {
                        "enum": [
                            "all",
                            "serial",
                            "usb",
                            "tcp",
                            "bluetooth"
                        ],
                        "type": "string",
                        "default": "all",
                        "description": "Scan type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "30s",
                        "description": "Scan timeout",
                        "name": "timeout",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device scan completed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "devices": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/definitions/service.DiscoveredDevice"
                                                    }
                                                },
                                                "devices_found": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Scan failed",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.AutoSetupRequest": {
            "type": "object",
            "required": [
                "branch_id"
            ],
            "properties": {
                "auto_connect": {
                    "type": "boolean"
                },
                "branch_id": {
                    "type": "string"
                },
                "device_filter": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "model.ConnectionType": {
            "type": "string",
            "enum": [
                "SERIAL",
                "USB",
                "TCP",
                "BLUETOOTH"
            ],
            "x-enum-varnames": [
                "ConnectionTypeSerial",
                "ConnectionTypeUSB",
                "ConnectionTypeTCP",
                "ConnectionTypeBluetooth"
            ]
        },
...
        "model.DeviceBrand": {
            "type": "string",
            "enum": [
                "EPSON",
                "STAR",
                "INGENICO",
                "PAX",
                "CITIZEN",
                "BIXOLON",
                "VERIFONE",
                "GENERIC"
            ],
            "x-enum-varnames": [
                "BrandEpson",
                "BrandStar",
                "BrandIngenico",
                "BrandPAX",
                "BrandCitizen",
                "BrandBixolon",
                "BrandVerifone",
                "BrandGeneric"
            ]
        },
...
        "model.DeviceType": {
            "type": "string",
            "enum": [
                "POS",
                "PRINTER",
                "SCANNER",
                "CASH_REGISTER",
                "CASH_DRAWER",
                "DISPLAY"
            ],
            "x-enum-varnames": [
                "DeviceTypePOS",
                "DeviceTypePrinter",
                "DeviceTypeScanner",
                "DeviceTypeCashRegister",
                "DeviceTypeCashDrawer",
                "DeviceTypeDisplay"
            ]
        },
...
        "service.AutoSetupResult": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "failed": {
                    "type": "integer"
                },
                "setup_devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.SetupDeviceResult"
                    }
                },
                "successfully_setup": {
                    "type": "integer"
                },
                "total_scanned": {
                    "type": "integer"
                }
            }
        },
...
        "service.DiscoveredDevice": {
            "type": "object",
            "properties": {
                "brand": {
                    "$ref": "#/definitions/model.DeviceBrand"
                },
                "capabilities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "confidence": {
                    "description": "0.0-1.0",
                    "type": "number"
                },
                "connection_info": {
                    "type": "object",
                    "additionalProperties": true
                },
                "connection_type": {
                    "$ref": "#/definitions/model.ConnectionType"
                },
                "device_type": {
                    "$ref": "#/definitions/model.DeviceType"
                },
                "location": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "serial_number": {
                    "type": "string"
                }
            }
        },
...
        "service.SetupDeviceResult": {
            "type": "object",
            "properties": {
                "brand": {
                    "$ref": "#/definitions/model.DeviceBrand"
                },
                "connection_type": {
                    "$ref": "#/definitions/model.ConnectionType"
                },
                "device_id": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "status": {
                    "description": "SUCCESS, FAILED",
                    "type": "string"
                }
            }
        },
...
        "utils.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "utils.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/utils.APIError"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}